# docs/openapi/fibness_api.yaml
openapi: 3.0.0
info:
  title: Fibness API
  version: "1.0.0"
servers:
  - url: 'http://10.4.41.146:3001'
    description: FIB server, UPC VPN required to connect
...
paths:
...
  /user/{id}/trainings:
    get:
      summary: Gives the trainings of a user
      description: Returns all the trainings of the user with the given id
      tags:
        - user
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the user to get the trainigs from
          schema:
            type: integer
            minimum: 1
      responses:
        200:
          description: Set of the ids, names and descriptions of the requested trainings
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    idelemento:
                      type: integer
                    nombre:
                      type: string
                    descripcion:
                      type: string
                    nlikes:
                      type: integer
                    ncomentarios:
                      type: integer
                example:
                  - idelemento: 28
                    nombre: TrainingTest
                    descripcion: DescriptionTest
                    nlikes: 20
                    ncomentarios: 2
                  - idelemento: 40
                    nombre: TrainingTest2
                    descripcion: DescriptionTest2
                    nlikes: 20
                    ncomentarios: 2
                required:
                - trainingSet
  /user/{id}/diets:
    get:
      summary: Gives the diets of a user
      description: Returns all the diets of the user with the given id
      tags:
        - user
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the user to get the diets from
          schema:
            type: integer
            minimum: 1
      responses:
        200:
          description: Set of the ids, names and descriptions of the requested diets
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    idelemento:
                      type: integer
                    nombre:
                      type: string
                    descripcion:
                      type: string
                    nlikes:
                      type: integer
                    ncomentarios:
                      type: integer
                example:
                  - idelemento: 23
                    nombre: DietTest
                    descripcion: DescriptionTest
                    nlikes: 20
                    ncomentarios: 2
                  - idelemento: 69
                    nombre: DietTest2
                    descripcion: DescriptionTest2
                    nlikes: 20
                    ncomentarios: 2
                required:
                - dietSet
  /user/{id}/routes:
    get:
      summary: Gives the routes of a user 
      description: Returns all the routes of the user with given id
      tags:
        - user
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the user, whose's routes we want to get
          schema:
            type: integer
            minimum: 1
      responses:
        200:
          description: Set of the ids, names and descriptions and others of the requested routes
          content:
            application/json:
              schema:
                type: object
                properties:
                  idelemento:
                    type: integer
                  nombre:
                    type: string
                  descripcion:
                    type: string
                  origen:
                    type: string
                  destino:
                    type: string
                  distancia:
                    type: string
                  nlikes:
                    type: integer
                  ncomentarios:
                    type: integer
                example:
                  - idelemento: 3
                    nombre: routeTest
                    descripcion: routeDescription
                    origen: 42.471615;2.271684
                    destino: 45.471615;5.271684
                    distancia: 60
                    nlikes: 20
                    ncomentarios: 2
                    
                  - idelemento: 4
                    nombre: routeTest2
                    descripcion: routeDescription2
                    origen: 43.471615;3.271684
                    destino: 46.471615;6.271684
                    distancia: 70
                    nlikes: 20
                    ncomentarios: 2
                    
                required:
                      - activitySet
...
  '/training':
    post:
      summary: Creates a training
      description: Creates a training with the given description and name associated with the given user
      tags:
        - training
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                  example: TrainingTest
                descripcion:
                  type: string
                  example: TrainingDescriptionTest
                idUser:
                  type: integer
                  description: id of the user
                  example: 2
              required:
                - nombre
                - descripcion
                - idUser
      responses:
        '201':
          description: Training created succesfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  idelemento:
                    type: integer
                    example: 12345
                required:
                  - idelemento
        '400':
          description: >-
            Some required parameters are missing or with an incorrect type
  '/training/{id}':
    delete:
      summary: Deletes a training by id
      tags:
        - training
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the training to delete
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Ok
    put:
      summary: Update a training by id
      tags:
        - training
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the training to update
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                  example: weekends training
                descripcion:
                  type: string
                  example: This training has vareity of exercices. 
              required:
                - name
                - description
      responses:
            '200':
              description: Ok
  /training/{idTraining}/activities:
    get:
      summary: Gives the activities(exercises or sports) of a certain training
      description: Returns all the activities of a given training id
      tags:
        - training
      parameters:
        - name: idTraining
          in: path
          required: true
          description: The id of the training
          schema:
            type: integer
            minimum: 1
      responses:
        200:
          description: Set of exercises or sports of a certain training
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    idActividad:
                      type: integer
                      description: id of the activity(sport or exercise)
                    nombre:
                      type: string
                     
                    descripcion:
                      type: string
                     
                    tiempoEjecucion:
                      type: integer
                      
                    numSets:
                      type: integer
        
                    numRepeticiones:
                      type: integer
                      
                    tiempoDescanso:
                      type: integer
                    posicion:
                      type: integer
                example:
                  - idactividad: 3
                    nombre: exerciseTest
                    descripcion: exerciseDescription
                    tiempoejecucion: 4
                    numsets: 3 
                    numrepeticiones: 2
                    tiempodescanso: 1
                    posicion: 3
                  - idactividad: 4
                    nombre: exerciseTest_2
                    descripcion: exerciseDescription_2
                    tiempoejecucion: 5
                    numsets: 4 
                    numrepeticiones: 1
                    tiempodescanso: 10  
                    posicion: 4
                required:
                  - activitySet
...
  '/diet':
    post:
      summary: Creates a diet
      description: Creates a diet with the given description and name associated with the given user
      tags:
        - diet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                  example: Hypertrophy diet
                descripcion:
                  type: string
                  example: Gain muscle
                idUser:
                  type: integer
                  description: id of the user
                  example: 2
              required:
                - nombre
                - descripcion
                - idUser
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  idelemento:
                    type: integer
                    example: 12345
                required:
                  - idelemento
        '400':
          description: >-
            Some required parameters are missing or with an incorrect type
  '/diet/{id}':
    delete:
      summary: Deletes a diet by id
      tags:
        - diet
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the diet to delete
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Ok
    put:
      summary: Update a diet by id
      tags:
        - diet
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the diet to update
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: weekends training
                description:
                  type: string
                  example: This training has vareity of exercices. 
              required:
                - name
                - description
      responses:
            200:
              description: Ok
  /diet/{idElemento}/{dia}:
    get:
      summary: Gives the meals of a certain day for a certain diet
      description: Returns all the meals of the diet with the given id for the specified day
      tags:
        - diet
      parameters:
        - name: idElemento
          in: path
          required: true
          description: The id of the diet to get the trainigs from
          schema:
            type: integer
            minimum: 1
        - name: dia
          in: path
          required: true
          description: day of the week
          schema:
            type: string
            enum: [lunes, martes, miercoles, jueves, viernes, sabado, domingo]
      responses:
        200:
          description: Set of the ids, names and timeToEat of the requested meals
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    idcomida:
                      type: integer
                    nombre:
                      type: string
                    horacomida:
                      type: string
                example:
                  - idcomida: 56
                    nombre: MealTest
                    horacomida: '10:00'
                  - idcomida: 100
                    nombre: MealTest2
                    horacomida: '18:00'
                required:
                - mealSet
...
  '/route':
    post:
      summary: Creates a route
      description: Creates a route with the given description and name associated with the given user
      tags:
        - route
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                  example: RouteTest
                descripcion:
                  type: string
                  example: RouteDescriptionTest
                idUser:
                  type: integer
                  description: id of the user
                  example: 2
                origen:
                  type: string
                  example: '42.471615;2.271684'
                destino:
                  type: string
                  example: '45.471615;5.271684'
                distancia:
                  type: string
                  example: '60' 
              required:
                - nombre
                - descripcion
                - idUser
                - origen
                - destino
                - distancia
      responses:
        '201':
          description: Training created succesfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  idElemento:
                    type: integer
                    example: 12345
                required:
                  - idElemento
        '400':
          description: >-
            Some required parameters are missing or with an incorrect type
  '/route/{id}':
    delete:
      summary: Deletes a route by id
      tags:
        - route
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the route to delete
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Ok
    put:
      summary: Update a route by id
      tags:
        - route
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the route to update
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                  example: RouteTest
                descripcion:
                  type: string
                  example: RouteDescriptionTest
                origen:
                  type: string
                  example: '42.471615;2.271684'
                destino:
                  type: string
                  example: '45.471615;5.271684'
                distancia:
                  type: string
                  example: '60' 
              required:
                - nombre
                - descripcion
                - origen
                - destino
                - distancia
      responses:
            '200':
              description: Ok
  '/comment/':
    post:
      summary: Creates a comment
      description: The user will create a comment on an elemetn
      tags:
        - comment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                idUser:
                  type: integer
                  example: 1
                  description: id of the user that will create the comment
                idElement:
                  type: integer
                  example: 2
                  description: id of the element OR COMMENT that will contain the comment
                text:
                  type: string
                  example: "nice"
                  description: text of the comment
              required:
                - idUser
                - idElement
                - text
      responses:
        '201':
          description: Comment created succesfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  idCom:
                    type: integer
                    example: 12345
                required:
                  - idCom
        '400':
          description: >-
            Some required parameters are missing or with an incorrect type
  '/comment/{idComment}':
    delete:
      summary: Deletes a comment 
      description: The comment will be eliminated
      tags:
        - comment
      parameters:
        - name: idComment
          in: path
          required: true
          description: The id of the comment to be deleted
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Comment removed
        '400':
          description: >-
            Some required parameters are missing or with an incorrect type
  '/comment/{idElemento}/comments':
    get:
      summary: Gets all the comments of an element
      description: Returns all the information of the comments of an element
      tags:
        - comment
      parameters:
        - name: idElemento
          in: path
          required: true
          description: The id of the element to get the comments from
          schema:
            type: integer
            minimum: 1
      responses:
        200:
          description: Information of the comments of the element
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    idcomentario:
                      type: integer
                      example: 5
                    idusuario:
                      type: integer
                      example: 3
                    fecha:
                      type: string
                      example: '2020-05-26T22:00:00.000Z'
                    texto:
                      type: string
                      example: "nice"
                    nlikes:
                      type: integer
                      example: 2